pyproject.toml
[tool.poetry]
name = "grpctunnel"
version = "0.1.0"
description = ""
authors = ["Monotosh Das <monotosh.das@videonetics.com>"]
readme = "README.md"
packages = [{include = "grpctunnel"}]

[tool.poetry.dependencies]
python = ">=3.8"
grpcio = "^1.59.3"
protobuf = "^4.25.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
grpcio-tools = "^1.59.3"
ruff = "^0.1.6"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
